Data-Engineering/1-Extract-Countries/requirements.txt
google-cloud-storage
opencage
pandas
wbgapi